training single texturize

+++ models/mesh_texturize.py
@@ -47,7 +47,7 @@ class TexturizeModel:
         #print("label: ",data['label'])
         #print("soft_label: ",data['soft_label'])
         input_edge_features = torch.from_numpy(data['edge_features']).float()
-        labels = torch.from_numpy(data['label']).long()
+        labels = torch.from_numpy(data['label']).float()
         # set inputs
         self.edge_features = input_edge_features.to(self.device).requires_grad_(self.is_train)
         self.labels = labels.to(self.device)
